Updating setup.py to implement python_requires attribute
Instead of requiring pickle5 for all users, I think this setup.py  is appropriate for users using python versions 3.8 and above, who already have the pickle dependency and can't install pickle5. Instead, I put a requirement of "jupyter" and enforce a python version of 3.8 or greater. This setup.py alone can be renamed and added to the main repository.